+++ example/lib/main.dart
@@ -67,6 +67,17 @@ class ReadyUI extends StatelessWidget {
         repoName: "tic-tac-no",
         locale: Locale("en"), // try this for english language
         // locale: Locale("ar"), // try this for arabic language
+        translations: const <ContributorsTranslation>[
+          ContributorsTranslation(
+            languageCode: 'es',
+            contributions: 'Contribuciones',
+            followers: 'Seguidores',
+            following: 'Siguiendo',
+            publicRepos: 'Reposiciones Públicas',
+            publicGists: 'Gists Públicos',
+            isRTL: false,
+          ),
+        ],
       ),
     );
   }
@@ -82,6 +93,17 @@ class CustomUI extends StatelessWidget {
         repoName: "tic-tac-no",
         locale: Locale("en"), // try this for english language
         // locale: Locale("ar"), // try this for arabic language
+        translations: const <ContributorsTranslation>[
+          ContributorsTranslation(
+            languageCode: 'es',
+            contributions: 'Contribuciones',
+            followers: 'Seguidores',
+            following: 'Siguiendo',
+            publicRepos: 'Reposiciones Públicas',
+            publicGists: 'Gists Públicos',
+            isRTL: false,
+          ),
+        ],
         builder: (
           BuildContext context,
           List<ContributorStatistics> contributorStatisticsList,