+++ lib/contributors.dart
@@ -3,5 +3,6 @@ library contributors;
 export 'package:contributors/src/ui/views/contributors/contributors_view.dart';
 export 'package:contributors/src/ui/views/contributors_builder/contributors_builder_view.dart';
 export 'package:contributors/src/ui/widgets/contributor_list_tile.dart';
+export 'package:contributors/src/app/utils/contributors_translations.dart';
 
 export 'package:github/github.dart';