+++ lib/src/ui/views/contributor/contributor_view.dart
@@ -1,3 +1,4 @@
+import 'package:contributors/src/app/utils/contributors_translations.dart';
 import 'package:contributors/src/app/utils/utils.dart';
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
@@ -21,6 +22,18 @@ class ContributorView extends StatelessWidget {
   /// [locale] is the [Locale] that you want the text to appear in.
   final Locale locale;
 
+  /// [translations] is the `List<ContributorsTranslation>` you want to use, we
+  /// have defaults but you can add custom ones here.
+  /// This overrides the default supported languages which means you should add
+  /// every language you want to support inside this list. The built-in
+  /// translations are factory methods of [ContributorsTranslation].
+  ///
+  /// Example:
+  ///
+  /// * ContributorsTranslation.en()
+  /// * ContributorsTranslation.ar()
+  final List<ContributorsTranslation> translations;
+
   /// `ContributorView` is the default contructor for [ContributorView]
   const ContributorView({
     Key key,
@@ -28,6 +41,7 @@ class ContributorView extends StatelessWidget {
     @required this.repoName,
     @required this.ownerName,
     @required this.locale,
+    @required this.translations,
   }) : super(key: key);
 
   @override
@@ -45,7 +59,7 @@ class ContributorView extends StatelessWidget {
         Widget child,
       ) {
         return Directionality(
-          textDirection: Utils.getCurrentTranslation(locale).isRTL
+          textDirection: Utils.getCurrentTranslation(locale, translations).isRTL
               ? TextDirection.rtl
               : TextDirection.ltr,
           child: Scaffold(
@@ -102,7 +116,7 @@ class ContributorView extends StatelessWidget {
                                   size: 25,
                                 ),
                                 title: Text(
-                                  '${Utils.getCurrentTranslation(locale).contributions} ($ownerName/$repoName)',
+                                  '${Utils.getCurrentTranslation(locale, translations).contributions} ($ownerName/$repoName)',
                                   style: const TextStyle(
                                     fontSize: 18,
                                     fontWeight: FontWeight.bold,
@@ -174,7 +188,9 @@ class ContributorView extends StatelessWidget {
                                   size: 25,
                                 ),
                                 title: Text(
-                                  Utils.getCurrentTranslation(locale).followers,
+                                  Utils.getCurrentTranslation(
+                                          locale, translations)
+                                      .followers,
                                   style: const TextStyle(
                                     fontSize: 18,
                                     fontWeight: FontWeight.bold,
@@ -201,7 +217,9 @@ class ContributorView extends StatelessWidget {
                                   size: 25,
                                 ),
                                 title: Text(
-                                  Utils.getCurrentTranslation(locale).following,
+                                  Utils.getCurrentTranslation(
+                                          locale, translations)
+                                      .following,
                                   style: const TextStyle(
                                     fontSize: 18,
                                     fontWeight: FontWeight.bold,
@@ -233,7 +251,8 @@ class ContributorView extends StatelessWidget {
                                   size: 25,
                                 ),
                                 title: Text(
-                                  Utils.getCurrentTranslation(locale)
+                                  Utils.getCurrentTranslation(
+                                          locale, translations)
                                       .publicRepos,
                                   style: const TextStyle(
                                     fontSize: 18,
@@ -261,7 +280,8 @@ class ContributorView extends StatelessWidget {
                                   size: 25,
                                 ),
                                 title: Text(
-                                  Utils.getCurrentTranslation(locale)
+                                  Utils.getCurrentTranslation(
+                                          locale, translations)
                                       .publicGists,
                                   style: const TextStyle(
                                     fontSize: 18,