+++ lib/src/ui/views/contributors_builder/contributors_builder_view.dart
@@ -1,3 +1,4 @@
+import 'package:contributors/src/app/utils/contributors_translations.dart';
 import 'package:contributors/src/app/utils/utils.dart';
 import 'package:flutter/material.dart';
 import 'package:github/github.dart';
@@ -25,6 +26,18 @@ class ContributorsBuilderView extends StatelessWidget {
   /// [locale] is the [Locale] that you want the text to appear in.
   final Locale locale;
 
+  /// [translations] is the `List<ContributorsTranslation>` you want to use, we
+  /// have defaults but you can add custom ones here.
+  /// This overrides the default supported languages which means you should add
+  /// every language you want to support inside this list. The built-in
+  /// translations are factory methods of [ContributorsTranslation].
+  ///
+  /// Example:
+  ///
+  /// * ContributorsTranslation.en()
+  /// * ContributorsTranslation.ar()
+  final List<ContributorsTranslation> translations;
+
   /// `ContributorsBuilderView` is the default contructor for [ContributorsBuilderView]
   const ContributorsBuilderView({
     Key key,
@@ -32,6 +45,7 @@ class ContributorsBuilderView extends StatelessWidget {
     @required this.ownerName,
     @required this.repoName,
     @required this.locale,
+    this.translations = const <ContributorsTranslation>[],
   }) : super(key: key);
 
   @override
@@ -50,7 +64,7 @@ class ContributorsBuilderView extends StatelessWidget {
         Widget child,
       ) {
         return Directionality(
-          textDirection: Utils.getCurrentTranslation(locale).isRTL
+          textDirection: Utils.getCurrentTranslation(locale, translations).isRTL
               ? TextDirection.rtl
               : TextDirection.ltr,
           child: Scaffold(