+++ lib/src/ui/widgets/contributor_list_tile.dart
@@ -1,3 +1,4 @@
+import 'package:contributors/src/app/utils/contributors_translations.dart';
 import 'package:contributors/src/app/utils/utils.dart';
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
@@ -21,6 +22,18 @@ class ContributorListTile extends StatelessWidget {
   /// [locale] is the [Locale] that you want the text to appear in.
   final Locale locale;
 
+  /// [translations] is the `List<ContributorsTranslation>` you want to use, we
+  /// have defaults but you can add custom ones here.
+  /// This overrides the default supported languages which means you should add
+  /// every language you want to support inside this list. The built-in
+  /// translations are factory methods of [ContributorsTranslation].
+  ///
+  /// Example:
+  ///
+  /// * ContributorsTranslation.en()
+  /// * ContributorsTranslation.ar()
+  final List<ContributorsTranslation> translations;
+
   /// `ContributorListTile` is the default contructor for [ContributorListTile]
   const ContributorListTile({
     Key key,
@@ -29,6 +42,7 @@ class ContributorListTile extends StatelessWidget {
     @required this.contributionsNumber,
     @required this.onTap,
     @required this.locale,
+    @required this.translations,
   }) : super(key: key);
 
   @override
@@ -39,7 +53,7 @@ class ContributorListTile extends StatelessWidget {
       ),
       title: Text(name),
       subtitle: Text(
-        '${Utils.getCurrentTranslation(locale).contributions}: $contributionsNumber',
+        '${Utils.getCurrentTranslation(locale, translations).contributions}: $contributionsNumber',
       ),
       onTap: onTap,
     );