+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## unreleased - unreleased
+
+- Add new way of adding custom translations
+
 ## 0.0.1
 
 - Basic functionality of getting contributors of a repository